widen liquid_pack_array, liquid_unpack_array, add block methods
this will allow us to switch ofdm from using repack_bytes to the new
block methods. this also tests the new methods and the widened original
ones.

this has the slight downside that unpack calls will now have to unpack
into a temp unsigned int and then copied into the symbol array if the
symbol array is of type unsigned char. this could be resolved by
creating new methods which have the types of the originals before this
patch with the original 8-bit symbol width limitation in place