make pack_array functions behave for "leftover" bytes
the previous logic was too conservative about what to do when the number
of bits in symbols exceeded array length in pack case and when byte
array was not long enough to fill symbols in unpack case. it now drops
excess bits in pack and shifts 0s into symbols in unpack.