makefile.in: Fix overloading of LDFLAGS and CFLAGS
The makefile.in used the variable LDFLAGS for both flags and libraries
to link. This will create problems as many flags need to appear before
the object files while the libraries to link need to appear after. A
similar problem stems from using CFLAGS for both include path
specifications and regular C compiler flags, as the CFLAGS are needed
for proper linking.

This patch splits LDFLAGS into LDFLAGS for flags only and LIBS for libs
only and splits CFLAGS into CFLAGS for compiler flags only and CPPFLAGS
for include path specifications only. Linking now changes LDFLAGS to
LIBS and adds CFLAGS and LDFLAGS at the beginning of the command. The
previously ignored @CPPFLAGS@ autoconf value is now also included in
CPPFLAGS.