Fix SONAME not set in shared library
A SONAME was introduced to libliquid and used in installation file names
however the shared library was always compiled with 'libliquid.so' as
the SONAME rather than 'libliquid.so.1' as was probably intended.

Fix that and introduce a SOVERSION variable to the makefile so that it
is defined in a single place only.